Add files via upload
This will allow one to create border around windows by defining the inner_border_width and inner_border_brightness that will make the pixels darker (or brighter) on the edge of the windows. This should work nicely with multicolored windows like dark headerbar but with white content. 

The two new options are:
inner-border-width = 1 (default 0)
inner-border-brightness = 0.8 (default 0.89)